Allow user to provide a vector of image handles, each to be used as a tile texture.

+++ examples/accessing_tiles.rs
@@ -88,7 +88,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
             size: tilemap_size,
             storage: tile_storage,
             map_type: tilemap_type,
-            texture: TilemapTexture(texture_handle),
+            texture: TilemapTexture::Single(texture_handle),
             tile_size,
             transform: get_tilemap_center_transform(&tilemap_size, &grid_size, 0.0),
             ..Default::default()
