+++ examples/colors.rs
@@ -83,7 +83,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
             grid_size,
             size: total_size,
             storage: tile_storage,
-            texture: TilemapTexture(texture_handle),
+            texture: TilemapTexture::Single(texture_handle),
             tile_size,
             map_type: TilemapType::Square {
                 diagonal_neighbors: false,
